tests/wpt/web-platform-tests/web-bundle/subresource-loading/script-relative-url-with-base.https.tentative.html

<!DOCTYPE html>
<title>Subresource loading using relative URLs in the 'resources' attribute with a base element</title>
<base href="../resources/wbn/">
<link rel="help" href="https://github.com/WICG/webpackage/blob/main/explainers/subresource-loading.md" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
  <script type="webbundle">
    {
      "source": "static-element.wbn",
      "resources": ["static-element/resources/script.js"]
    }
  </script>
  <script id="script" src="static-element/resources/script.js"></script>

  <script type="webbundle">
    {
      "source": "dynamic1.wbn",
      "scopes": ["dynamic/resource"]
    }
  </script>

  <script>
    setup(() => {
      assert_true(HTMLScriptElement.supports('webbundle'));
    });
  </script>
  <script src="/web-bundle/subresource-loading/resources/relative-url-with-base-test.js"></script>
</body>
